libs/sywi_duckdb/pyproject.toml
[project]
name = "sywi-duckdb"
version = "0.1.0"
description = "DuckDB wrapper with auto-configured SYWI DuckLake connections"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sywi_duckdb"]
